# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: H Fans API - World App Mini App creator platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /messages/conversations/{userId}:
    get:
      operationId: getMessages
      tags: [messages]
      summary: Get messages in conversation
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageListResponse"

    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageRequest"
      responses:
        "201":
          description: Sent message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
  /messages/mass-dm:
    post:
      operationId: sendMassDM
      tags: [messages]
      summary: Send mass DM to subscribers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MassDMRequest"
      responses:
        "200":
          description: Mass DM sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # LIVE STREAMING
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]

    # AUTH
...
    Message:
      type: object
      properties:
        id:
          type: string
        senderId:
          type: string
        receiverId:
          type: string
        text:
          type: string
        mediaUrl:
          type: string
        mediaType:
          type: string
          enum: [image, video]
        isPpv:
          type: boolean
        ppvPriceWld:
          type: string
        isUnlocked:
          type: boolean
        isRead:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, senderId, receiverId, createdAt]
    SendMessageRequest:
      type: object
      properties:
        text:
          type: string
        mediaId:
          type: string
        isPpv:
          type: boolean
        ppvPriceWld:
          type: string
    MessageListResponse:
      type: object
      properties:
        messages:
          type: array
          items:
            $ref: "#/components/schemas/Message"
        total:
          type: integer
        page:
          type: integer
      required: [messages, total, page]
...
    MassDMRequest:
      type: object
      properties:
        text:
          type: string
        mediaId:
          type: string
        isPpv:
          type: boolean
        ppvPriceWld:
          type: string
        listId:
          type: string
      required: [text]

    # LIVESTREAM
...